Security Enhancement
Switched from ROLE_ to PERM_ based security for fine grained access control.
All Roles and Permissions are now in the database.

A Permission is a fine grained ACL which you assign using
@Secured({"PERM_VIEW_USER"})

Permissions can only be assigned to a Role.

Roles can be assigned to users which contain permissions.
It is a way of grouping permissions.